refactor uploader
